+++ internal/cmd/dataimports/makeplanetscalereplica.go
@@ -77,7 +77,7 @@ func MakePlanetScaleReplicaCmd(ch *cmdutil.Helper) *cobra.Command {
 			}
 			end()
 			ch.Printer.Printf("Successfully switch PlanetScale database %s to Replica.\n", printer.BoldBlue(flags.name))
-			ch.Printer.PrintDataImport(*dataImport)
+			PrintDataImport(ch.Printer, *dataImport)
 			return nil
 		},
 	}
